Refactor project into layered structure
