+++ phone_spam_checker/domain/models.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+@dataclass
+class PhoneCheckResult:
+    """Result of phone number verification."""
+    phone_number: str
+    status: str
+    details: str = ""
